feat: updating to lore v2 to allow better formatting and searching

+++ xivexport/__main__.py
@@ -5,6 +5,7 @@ from pprint import pp
 from dotenv import load_dotenv
 from src.xivexport import adapter, search, xivclient, model
 from itertools import batched
+import json
 
 # Load environment variables from .env file
 load_dotenv()
@@ -20,6 +21,7 @@ VERSION_NAME = os.listdir(INPUT_PATH)[-1]
 DATA_PATH = f"{INPUT_PATH}\\{VERSION_NAME}\\exd"
 OUTPUT_PATH = f"{ROOT_PATH}\\_dumps\\output"
 OUTPUT_FILE = f"{OUTPUT_PATH}\\dump.txt"
+OUTPUT_FILE_JSON = f"{OUTPUT_PATH}\\dump.json"
 
 def delete_dump_file():        
     if os.path.exists(OUTPUT_FILE):
@@ -28,17 +30,30 @@ def delete_dump_file():
     else:
         print(f"File '{OUTPUT_FILE}' does not exist.")
 
+    if os.path.exists(OUTPUT_FILE_JSON):
+        os.remove(OUTPUT_FILE_JSON)
+        print(f"File '{OUTPUT_FILE_JSON}' deleted successfully.")
+    else:
+        print(f"File '{OUTPUT_FILE_JSON}' does not exist.")
+
 def dump_docs(docs):
     """Outputs the search items as plaintext to a txt file. Mainly for debugging purposes."""
 
     print("Outputting text files...")
-    with open(OUTPUT_FILE, "a", encoding="UTF-8") as fh:
+    with open(OUTPUT_FILE, "a", encoding="UTF-8") as fh:        
         for doc in docs:
-            fh.write(_serialize(doc))
+            fh.write('\n')
+            fh.write(doc.as_plain_text())
+    
+    data = []
+    try :
+        with open(OUTPUT_FILE_JSON, "r", encoding="UTF-8") as fh:
+            data = json.load(fh)
+    except Exception:
+        data = []
 
-
-def _serialize(doc: model.SearchItem):
-    return doc.as_plain_text()
+    with open(OUTPUT_FILE_JSON, "w", encoding="UTF-8") as fh:
+        json.dump(data + [x.model_dump() for x in docs], fh, indent=2)
 
 def connect():
     search.ClientManager.connect(API_USER, API_SECRET)
@@ -48,29 +63,34 @@ def close():
     search.ClientManager.close()
     xivclient.XivApiClientManager.close()
 
-def save_batch(docs: list[model.SearchItem]):
+def save_batch(docs: list[model.SearchItem], remote_save=True):
     dump_docs(docs)
     print(f"inserting records...")
-    search.ClientManager.upload_docs([d.model_dump() for d in docs])
+    if remote_save:
+        search.ClientManager.upload_docs([d.model_dump() for d in docs])
 
-def clear_data():
+def clear_data(is_test):
     # pass
     delete_dump_file()
-    print("Truncating records...")
-    search.ClientManager.truncate()
+    
+    if not is_test:
+        print("Truncating records...")
+        search.ClientManager.truncate()
 
 def main():
     """Entry point for the xivexport application."""
+    debug = True
     connect()
-    clear_data()
+
+    clear_data(debug)
 
     try:
-        adapters = adapter.__all__
+        adapters = [adapter.FishAdapter]  #adapter.__all__
 
         for adp in adapters:
             print(f"Loading docs for {adp.__name__}...")
-            for docs in batched(adp.get_all(), 1000):
-                save_batch(docs)
+            for docs in batched(adp.get_all(), 10):
+                save_batch(docs, not debug)
             print(f"*******************************\nDone with {adp.__name__}!\n*******************************")
     finally:
         print("Done!")
